third_party/WebKit/LayoutTests/platform/mac-mac10.10/virtual/outofblink-cors-http/http/tests/fetch/serviceworker/thorough/redirect-expected.txt

This is a testharness.js-based test.
PASS Startup 
PASS redirect 
PASS executeTest-0 
PASS executeTest-1 
PASS executeTest-2 
PASS executeTest-3 
PASS executeTest-4 
PASS executeTest-5 
PASS executeTest-6 
PASS executeTest-7 
PASS executeTest-8 
PASS executeTest-9 
PASS executeTest-10 
PASS executeTest-11 
PASS executeTest-12 
PASS executeTest-13 
PASS executeTest-14 
PASS executeTest-15 
PASS executeTest-16 
PASS executeTest-17 
PASS executeTest-18 
FAIL executeTest-19 assert_equals: http://127.0.0.1:8000/serviceworker/resources/redirect.php?Redirect=http%3A%2F%2Flocalhost%3A8000%2Fserviceworker%2Fresources%2Ffetch-access-control.php%3F%26ACAOrigin%3D*&mode=cors&credentials=same-origin&method=GET should be resolved expected "resolved" but got "rejected"
PASS executeTest-20 
FAIL executeTest-21 assert_equals: http://127.0.0.1:8000/serviceworker/resources/redirect.php?Redirect=http%3A%2F%2Flocalhost%3A8000%2Fserviceworker%2Fresources%2Ffetch-access-control.php%3F%26ACAOrigin%3D*%26ACAMethods%3DPUT&mode=cors&credentials=same-origin&method=PUT should be resolved expected "resolved" but got "rejected"
FAIL executeTest-22 assert_equals: http://127.0.0.1:8000/serviceworker/resources/redirect.php?Redirect=http%3A%2F%2Flocalhost%3A8000%2Fserviceworker%2Fresources%2Ffetch-access-control.php%3F%26ACAOrigin%3D*&mode=cors&credentials=same-origin&method=POST&Status=301 should be resolved expected "resolved" but got "rejected"
FAIL executeTest-23 assert_equals: http://127.0.0.1:8000/serviceworker/resources/redirect.php?Redirect=http%3A%2F%2Flocalhost%3A8000%2Fserviceworker%2Fresources%2Ffetch-access-control.php%3F%26ACAOrigin%3D*&mode=cors&credentials=same-origin&method=POST&Status=302 should be resolved expected "resolved" but got "rejected"
FAIL executeTest-24 assert_equals: http://127.0.0.1:8000/serviceworker/resources/redirect.php?Redirect=http%3A%2F%2Flocalhost%3A8000%2Fserviceworker%2Fresources%2Ffetch-access-control.php%3F%26ACAOrigin%3D*&mode=cors&credentials=same-origin&method=POST&Status=303 should be resolved expected "resolved" but got "rejected"
FAIL executeTest-25 assert_equals: http://127.0.0.1:8000/serviceworker/resources/redirect.php?Redirect=http%3A%2F%2Flocalhost%3A8000%2Fserviceworker%2Fresources%2Ffetch-access-control.php%3F%26ACAOrigin%3D*&mode=cors&credentials=same-origin&method=POST&Status=307 should be resolved expected "resolved" but got "rejected"
FAIL executeTest-26 assert_equals: http://127.0.0.1:8000/serviceworker/resources/redirect.php?Redirect=http%3A%2F%2Flocalhost%3A8000%2Fserviceworker%2Fresources%2Ffetch-access-control.php%3FACAOrigin%3D*%26ACAHeaders%3Dx-serviceworker-test&mode=cors&credentials=same-origin&method=GET&headers=CUSTOM should be resolved expected "resolved" but got "rejected"
PASS executeTest-27 
PASS executeTest-28 
FAIL executeTest-29 assert_equals: http://127.0.0.1:8000/serviceworker/resources/redirect.php?Redirect=http%3A%2F%2Flocalhost%3A8000%2Fserviceworker%2Fresources%2Ffetch-access-control.php%3F%26ACAOrigin%3Dhttp%3A%2F%2F127.0.0.1%3A8000%26ACEHeaders%3DContent-Length%2C%20X-ServiceWorker-ServerHeader&mode=cors&credentials=same-origin&method=GET should be resolved expected "resolved" but got "rejected"
PASS executeTest-30 
PASS executeTest-31 
PASS executeTest-32 
PASS executeTest-33 
FAIL executeTest-34 assert_equals: http://localhost:8000/serviceworker/resources/redirect.php?Redirect=http%3A%2F%2F127.0.0.1%3A8000%2Fserviceworker%2Fresources%2Ffetch-access-control.php%3FACAOrigin%3D*&mode=cors&credentials=same-origin&method=GET&ACAOrigin=* should be resolved expected "resolved" but got "rejected"
FAIL executeTest-35 assert_equals: http://localhost:8000/serviceworker/resources/redirect.php?Redirect=http%3A%2F%2F127.0.0.1%3A8000%2Fserviceworker%2Fresources%2Ffetch-access-control.php%3FACAOrigin%3D*&mode=cors&credentials=same-origin&method=post&ACAOrigin=*&Status=301 should be resolved expected "resolved" but got "rejected"
FAIL executeTest-36 assert_equals: http://localhost:8000/serviceworker/resources/redirect.php?Redirect=http%3A%2F%2F127.0.0.1%3A8000%2Fserviceworker%2Fresources%2Ffetch-access-control.php%3FACAOrigin%3D*&mode=cors&credentials=same-origin&method=post&ACAOrigin=*&Status=302 should be resolved expected "resolved" but got "rejected"
FAIL executeTest-37 assert_equals: http://localhost:8000/serviceworker/resources/redirect.php?Redirect=http%3A%2F%2F127.0.0.1%3A8000%2Fserviceworker%2Fresources%2Ffetch-access-control.php%3FACAOrigin%3D*&mode=cors&credentials=same-origin&method=post&ACAOrigin=*&Status=303 should be resolved expected "resolved" but got "rejected"
FAIL executeTest-38 assert_equals: http://localhost:8000/serviceworker/resources/redirect.php?Redirect=http%3A%2F%2F127.0.0.1%3A8000%2Fserviceworker%2Fresources%2Ffetch-access-control.php%3FACAOrigin%3D*&mode=cors&credentials=same-origin&method=post&ACAOrigin=*&Status=307 should be resolved expected "resolved" but got "rejected"
FAIL executeTest-39 assert_equals: http://localhost:8000/serviceworker/resources/redirect.php?Redirect=http%3A%2F%2F127.0.0.1%3A8000%2Fserviceworker%2Fresources%2Ffetch-access-control.php%3FACAOrigin%3D*%26ACAMethods%3DPUT&mode=cors&credentials=same-origin&method=PUT&ACAOrigin=*&ACAMethods=PUT should be resolved expected "resolved" but got "rejected"
FAIL executeTest-40 assert_equals: http://localhost:8000/serviceworker/resources/redirect.php?Redirect=http%3A%2F%2F127.0.0.1%3A8000%2Fserviceworker%2Fresources%2Ffetch-access-control.php%3FACAOrigin%3D*%26ACAHeaders%3Dx-serviceworker-test&mode=cors&credentials=same-origin&method=GET&headers=CUSTOM&ACAOrigin=*&ACAHeaders=x-serviceworker-test should be resolved expected "resolved" but got "rejected"
PASS executeTest-41 
PASS executeTest-42 
PASS executeTest-43 
FAIL executeTest-44 assert_equals: http://localhost:8000/serviceworker/resources/redirect.php?Redirect=http%3A%2F%2Flocalhost%3A8000%2Fserviceworker%2Fresources%2Ffetch-access-control.php%3FACAOrigin%3D*&mode=cors&credentials=same-origin&method=GET&ACAOrigin=* should be resolved expected "resolved" but got "rejected"
FAIL executeTest-45 assert_equals: http://localhost:8000/serviceworker/resources/redirect.php?Redirect=http%3A%2F%2Flocalhost%3A8000%2Fserviceworker%2Fresources%2Ffetch-access-control.php%3FACAOrigin%3Dhttp%3A%2F%2F127.0.0.1%3A8000&mode=cors&credentials=same-origin&method=GET&ACAOrigin=* should be resolved expected "resolved" but got "rejected"
FAIL executeTest-46 assert_equals: http://localhost:8000/serviceworker/resources/redirect.php?Redirect=http%3A%2F%2Flocalhost%3A8000%2Fserviceworker%2Fresources%2Ffetch-access-control.php%3FACAOrigin%3D*&mode=cors&credentials=same-origin&method=GET&ACAOrigin=http://127.0.0.1:8000 should be resolved expected "resolved" but got "rejected"
FAIL executeTest-47 assert_equals: http://localhost:8000/serviceworker/resources/redirect.php?Redirect=http%3A%2F%2Flocalhost%3A8000%2Fserviceworker%2Fresources%2Ffetch-access-control.php%3FACAOrigin%3Dhttp%3A%2F%2F127.0.0.1%3A8000&mode=cors&credentials=same-origin&method=GET&ACAOrigin=http://127.0.0.1:8000 should be resolved expected "resolved" but got "rejected"
FAIL executeTest-48 assert_equals: http://localhost:8000/serviceworker/resources/redirect.php?Redirect=http%3A%2F%2Flocalhost%3A8000%2Fserviceworker%2Fresources%2Ffetch-access-control.php%3FACAOrigin%3D*&mode=cors&credentials=same-origin&method=POST&ACAOrigin=*&Status=301 should be resolved expected "resolved" but got "rejected"
FAIL executeTest-49 assert_equals: http://localhost:8000/serviceworker/resources/redirect.php?Redirect=http%3A%2F%2Flocalhost%3A8000%2Fserviceworker%2Fresources%2Ffetch-access-control.php%3FACAOrigin%3D*&mode=cors&credentials=same-origin&method=POST&ACAOrigin=*&Status=302 should be resolved expected "resolved" but got "rejected"
FAIL executeTest-50 assert_equals: http://localhost:8000/serviceworker/resources/redirect.php?Redirect=http%3A%2F%2Flocalhost%3A8000%2Fserviceworker%2Fresources%2Ffetch-access-control.php%3FACAOrigin%3D*&mode=cors&credentials=same-origin&method=POST&ACAOrigin=*&Status=303 should be resolved expected "resolved" but got "rejected"
FAIL executeTest-51 assert_equals: http://localhost:8000/serviceworker/resources/redirect.php?Redirect=http%3A%2F%2Flocalhost%3A8000%2Fserviceworker%2Fresources%2Ffetch-access-control.php%3FACAOrigin%3D*&mode=cors&credentials=same-origin&method=POST&ACAOrigin=*&Status=307 should be resolved expected "resolved" but got "rejected"
FAIL executeTest-52 assert_equals: http://localhost:8000/serviceworker/resources/redirect.php?Redirect=http%3A%2F%2Flocalhost%3A8000%2Fserviceworker%2Fresources%2Ffetch-access-control.php%3FACAOrigin%3D*%26ACEHeaders%3DX-ServiceWorker-ServerHeader&mode=cors&credentials=same-origin&method=GET&ACAOrigin=* should be resolved expected "resolved" but got "rejected"
FAIL executeTest-53 assert_equals: http://localhost:8000/serviceworker/resources/redirect.php?Redirect=http%3A%2F%2Flocalhost%3A8000%2Fserviceworker%2Fresources%2Ffetch-access-control.php%3FACAOrigin%3D*%26ACAMethods%3DPUT&mode=cors&credentials=same-origin&method=PUT&ACAOrigin=*&ACAMethods=PUT should be resolved expected "resolved" but got "rejected"
FAIL executeTest-54 assert_equals: http://localhost:8000/serviceworker/resources/redirect.php?Redirect=http%3A%2F%2Flocalhost%3A8000%2Fserviceworker%2Fresources%2Ffetch-access-control.php%3FACAOrigin%3Dhttp%3A%2F%2F127.0.0.1%3A8000%26ACAHeaders%3Dx-serviceworker-test&mode=cors&credentials=same-origin&method=GET&headers=CUSTOM&ACAOrigin=http://127.0.0.1:8000&ACAHeaders=x-serviceworker-test should be resolved expected "resolved" but got "rejected"
Harness: the test ran to completion.
